requirements.txt
torch==2.4.0+cu118
torchvision==0.19.0+cu118
numpy==1.26.4
opencv-python==4.5.5.64
matplotlib==3.10.1
tqdm==4.67.1
lpips==0.1.4
scikit-image==0.19.3
Pillow==10.2.0
pytorch-msssim==1.0.0
einops==0.8.1
ptflops==0.7.4
